homework的size类型改为long,教师安排表对学生课表增加双向关联。

+++ easyclass-domain/src/main/java/edu/vero/easyclass/domain/Homework.java
@@ -26,11 +26,12 @@ public class Homework
 
     private Date deadline;
 
-    private int size;
+    private Long size;
 
     private boolean isSubmitted;
 
     private ClassSchedule schedule;
+
 
     @ManyToOne
     @JoinColumn(name="scheduleId")
@@ -115,12 +116,12 @@ public class Homework
         this.deadline = deadline;
     }
 
-    public int getSize()
+    public Long getSize()
     {
         return size;
     }
 
-    public void setSize(int size)
+    public void setSize(Long size)
     {
         this.size = size;
     }
